+++ client/web/src/insights/pages/creation/search-insight/components/search-insight-creation-content/SearchInsightCreationContent.tsx
@@ -5,7 +5,7 @@ import { noop } from 'rxjs'
 import { Settings } from '@sourcegraph/shared/src/settings/settings'
 
 import { useField } from '../../../../../components/form/hooks/useField'
-import { SubmissionErrors, useForm } from '../../../../../components/form/hooks/useForm'
+import { FormChangeEvent, SubmissionErrors, useForm } from '../../../../../components/form/hooks/useForm'
 import { useTitleValidator } from '../../../../../components/form/hooks/useTitleValidator'
 import { Organization } from '../../../../../components/visibility-picker/VisibilityPicker'
 import { InsightTypePrefix } from '../../../../../core/types'
@@ -51,6 +51,8 @@ export interface SearchInsightCreationContentProps {
     onSubmit: (values: CreateInsightFormFields) => SubmissionErrors | Promise<SubmissionErrors> | void
     /** Cancel handler. */
     onCancel?: () => void
+    /** Change handlers is called every time when user changed any field within the form. */
+    onChange?: (event: FormChangeEvent<CreateInsightFormFields>) => void
 }
 
 export const SearchInsightCreationContent: React.FunctionComponent<SearchInsightCreationContentProps> = props => {
@@ -63,13 +65,15 @@ export const SearchInsightCreationContent: React.FunctionComponent<SearchInsight
         dataTestId,
         onSubmit,
         onCancel = noop,
+        onChange = noop,
     } = props
 
     const isEditMode = mode === 'edit'
 
-    const { formAPI, ref, handleSubmit } = useForm<CreateInsightFormFields>({
+    const { values, formAPI, ref, handleSubmit } = useForm<CreateInsightFormFields>({
         initialValues: initialValue,
         onSubmit,
+        onChange,
         touched: isEditMode,
     })
 
@@ -91,6 +95,18 @@ export const SearchInsightCreationContent: React.FunctionComponent<SearchInsight
         series,
     })
 
+    const handleFormReset = (): void => {
+        // TODO [VK] Change useForm API in order to implement form.reset method.
+        title.input.onChange('')
+        repositories.input.onChange('')
+        // Focus first element of the form
+        repositories.input.ref.current?.focus()
+        visibility.input.onChange('personal')
+        series.input.onChange([createDefaultEditSeries({ edit: true })])
+        stepValue.input.onChange('2')
+        step.input.onChange('months')
+    }
+
     const validEditSeries = editSeries.filter(series => series.valid)
 
     // If some fields that needed to run live preview  are invalid
@@ -99,6 +115,11 @@ export const SearchInsightCreationContent: React.FunctionComponent<SearchInsight
         (repositories.meta.validState === 'VALID' || repositories.meta.validState === 'CHECKING') &&
         (series.meta.validState === 'VALID' || validEditSeries.length) &&
         stepValue.meta.validState === 'VALID'
+
+    const hasFilledValue =
+        values.series?.some(line => line.name !== '' || line.query !== '') ||
+        values.repositories !== '' ||
+        values.title !== ''
 
     return (
         <div data-testid={dataTestId} className={classnames(styles.content, className)}>
@@ -117,12 +138,14 @@ export const SearchInsightCreationContent: React.FunctionComponent<SearchInsight
                 series={series}
                 step={step}
                 stepValue={stepValue}
+                isFormClearActive={hasFilledValue}
                 onSeriesLiveChange={listen}
                 onCancel={onCancel}
                 onEditSeriesRequest={editRequest}
                 onEditSeriesCancel={cancelEdit}
                 onEditSeriesCommit={editCommit}
                 onSeriesRemove={deleteSeries}
+                onFormReset={handleFormReset}
             />
 
             <SearchInsightLivePreview